Security Update 2
- User Roles Service
- User Permissions Service
- Roles Permissions Service
- cascade Role - Permissions deletion
- Controllers will be added later

+++ src/main/java/com/example/security/rolesPermissions/RolesPermissionsService.java
@@ -1,15 +1,23 @@
 package com.example.security.rolesPermissions;
 
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.example.aspect.Exceptions;
+import com.example.security.permissions.Permissions;
+import com.example.security.roles.Roles;
+import com.example.security.roles.RolesService;
 
 @Service
 public class RolesPermissionsService {
 
 	@Autowired
 	RolePermissionRepository rolePermissionRepository ; 
+	
+	@Autowired 
+	private RolesService rolesService ; 
 	
 	public void addRolePermission(RolePermission rolePermission) {
 		if(this.rolePermissionRepository.findAll().contains(rolePermission)) {
@@ -18,8 +26,25 @@ public class RolesPermissionsService {
 		this.rolePermissionRepository.save(rolePermission);
 	}
 	
-	public void deleteRolePermission(RolePermission rolePermission) {
-		this.rolePermissionRepository.delete(rolePermission);
+	
+	//
+	public void deletePermission(Permissions permission) {
+		List<RolePermission> rolePermissionList = this.rolePermissionRepository.findAll() ;
+		for(RolePermission rolePermission : rolePermissionList ) {
+			if(rolePermission.getPermission().getPermissionName().equalsIgnoreCase(permission.getPermissionName())) {
+				this.rolesService.revokePermissionFromRoles(rolePermission.getPermission(),rolePermission.getRole());
+				this.rolePermissionRepository.delete(rolePermission);
+			}
+		}
+	}
+	
+	public void deleteRole(Roles role ) {
+		List<RolePermission> rolePermissionList = this.rolePermissionRepository.findAll(); 
+		for(RolePermission rolePermission : rolePermissionList) {
+			if(rolePermission.getRole().getRoleName().equalsIgnoreCase(role.getRoleName())) {
+				this.rolePermissionRepository.delete(rolePermission);
+			}
+		}
 	}
 	
 	public void updateRolePermission(RolePermission rolePermission) {
